change int to size_t in accept_connection
git-svn-id: http://svn.ratbox.org/svnroot/trunk@388 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/listener.c
@@ -320,7 +320,7 @@ void accept_connection(struct Listener* listener)
 {
   static time_t      last_oper_notice = 0;
   struct sockaddr_in addr;
-  int                addrlen = sizeof(struct sockaddr_in);
+  size_t             addrlen = sizeof(struct sockaddr_in);
   int                fd;
 
   assert(0 != listener);
